+++ INTERX/config/init.go
@@ -18,10 +18,13 @@ func getGetMethods() []string {
 		QueryDeposits,
 		QueryStatus,
 		QueryValidators,
+		QueryValidatorInfos,
 		QueryBlocks,
 		QueryBlockByHeightOrHash,
 		QueryBlockTransactions,
 		QueryTransactionResult,
+		QueryProposals,
+		QueryProposal,
 	}
 }