+++ INTERX/gateway/main.go
@@ -8,11 +8,14 @@ import (
 
 	"github.com/KiraCore/sekai/INTERX/config"
 	"github.com/KiraCore/sekai/INTERX/database"
+	"github.com/KiraCore/sekai/INTERX/functions"
 	cosmosAuth "github.com/KiraCore/sekai/INTERX/proto-gen/cosmos/auth"
 	cosmosBank "github.com/KiraCore/sekai/INTERX/proto-gen/cosmos/bank"
 	kiraGov "github.com/KiraCore/sekai/INTERX/proto-gen/kira/gov"
+	kiraSlashing "github.com/KiraCore/sekai/INTERX/proto-gen/kira/slashing"
 	kiraStaking "github.com/KiraCore/sekai/INTERX/proto-gen/kira/staking"
 	"github.com/KiraCore/sekai/INTERX/tasks"
+	functionmeta "github.com/KiraCore/sekai/function_meta"
 	"github.com/gorilla/mux"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/rakyll/statik/fs"
@@ -78,12 +81,20 @@ func GetGrpcServeMux(grpcAddr string) (*runtime.ServeMux, error) {
 		return nil, fmt.Errorf("failed to register gateway: %w", err)
 	}
 
+	err = kiraSlashing.RegisterQueryHandler(context.Background(), gwCosmosmux, conn)
+	if err != nil {
+		return nil, fmt.Errorf("failed to register gateway: %w", err)
+	}
+
 	return gwCosmosmux, nil
 }
 
 // Run runs the gRPC-Gateway, dialling the provided address.
 func Run(configFilePath string, log grpclog.LoggerV2) error {
 	config.LoadConfig(configFilePath)
+	functions.RegisterInterxFunctions()
+	functionmeta.RegisterStdMsgs()
+
 	database.LoadBlockDbDriver()
 	database.LoadFaucetDbDriver()
 	database.LoadReferenceDbDriver()