+++ x/staking/abci.go
@@ -6,66 +6,7 @@ import (
 	abci "github.com/tendermint/tendermint/abci/types"
 )
 
+// EndBlocker called every block, update validator set
 func EndBlocker(ctx sdk.Context, k keeper.Keeper) []abci.ValidatorUpdate {
-	var valUpdate []abci.ValidatorUpdate
-
-	valSet := k.GetPendingValidatorSet(ctx)
-	for _, val := range valSet {
-		k.AddValidator(ctx, val)
-		k.AfterValidatorCreated(ctx, val.ValKey)
-
-		consPk, err := val.TmConsPubKey()
-		if err != nil {
-			panic(err)
-		}
-
-		valUpdate = append(valUpdate, abci.ValidatorUpdate{
-			Power:  1,
-			PubKey: consPk,
-		})
-
-		k.RemovePendingValidator(ctx, val)
-	}
-
-	// Remove validators from the set, paused or inactivated.
-	removeVals := k.GetRemovingValidatorSet(ctx)
-	for _, val := range removeVals {
-		validator, err := k.GetValidator(ctx, val)
-		if err != nil {
-			panic("validator not found")
-		}
-
-		consPk, err := validator.TmConsPubKey()
-		if err != nil {
-			panic(err)
-		}
-
-		valUpdate = append(valUpdate, abci.ValidatorUpdate{
-			Power:  0,
-			PubKey: consPk,
-		})
-		k.RemoveRemovingValidator(ctx, validator)
-	}
-
-	// Remove validators from the set, paused or inactivated.
-	reactivateVals := k.GetReactivatingValidatorSet(ctx)
-	for _, val := range reactivateVals {
-		validator, err := k.GetValidator(ctx, val)
-		if err != nil {
-			panic("validator not found")
-		}
-
-		consPk, err := validator.TmConsPubKey()
-		if err != nil {
-			panic(err)
-		}
-
-		valUpdate = append(valUpdate, abci.ValidatorUpdate{
-			Power:  1,
-			PubKey: consPk,
-		})
-		k.RemoveRemovingValidator(ctx, validator)
-	}
-
-	return valUpdate
+	return k.BlockValidatorUpdates(ctx)
 }