+++ x/staking/types/expected_keepers.go
@@ -7,6 +7,7 @@ import (
 
 // StakingHooks event hooks for staking validator object (noalias)
 type StakingHooks interface {
+	AfterValidatorJoined(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress)  // Must be called when a validator is joined
 	AfterValidatorCreated(ctx sdk.Context, valAddr sdk.ValAddress)                           // Must be called when a validator is created
 	BeforeValidatorModified(ctx sdk.Context, valAddr sdk.ValAddress)                         // Must be called when a validator's state changes
 	AfterValidatorRemoved(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) // Must be called when a validator is deleted
@@ -16,4 +17,7 @@ type StakingHooks interface {
 type GovKeeper interface {
 	// returns network properties
 	GetNetworkProperties(sdk.Context) *customgovtypes.NetworkProperties
+	// GetNetworkActorsByAbsoluteWhitelistPermission returns all actors that have a specific whitelist permission,
+	// it does not matter if it is by role or by individual permission.
+	GetNetworkActorsByAbsoluteWhitelistPermission(ctx sdk.Context, perm customgovtypes.PermValue) []customgovtypes.NetworkActor
 }